pyproject.toml
[project]
name = "xprinter-gui"
version = "0.1.0"
description = "a python api and GUI to manage and operate xprinter thermal printers "
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pillow>=12.2.0",
    "qrcode>=8.2",
    "customtkinter>=5.2.0",
]
